Improve typed i18n infrastructure + explicit namespace
